// packages/formdescriptioneditors/frontend/sirius-components-formdescriptioneditors/src/SplitButtonWidget.type.ts
/*******************************************************************************
 * Copyright (c) 2024 Obeo.
 * This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v2.0
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Obeo - initial API and implementation
 *******************************************************************************/

export interface SplitButtonWidgetState {
  actionsButtonLabel: string[];
  actionsImageURL: string[];
  actionsIsValidImage: boolean[];
  selected: boolean;
  open: boolean;
  selectedIndex: number;
  message: string;
}